moved map marker data into a custom hook
